+++ tsconfig.json
@@ -12,6 +12,9 @@
     "resolveJsonModule": true,
     "useUnknownInCatchVariables": false,
   },
+  "include": [
+    "src/**/*"
+  ],
   "exclude": [
     "./__mocks__/**/*",
     "./__tests__/**/*",
